requirements.txt
httpx==0.27.2
langchain==0.0.352
openai==1.3.5
python-multipart==0.0.6
fastapi==0.104.1
uvicorn==0.24.0
pymupdf==1.24.0
python-docx==1.0.1
faiss-cpu==1.7.4
unstructured==0.10.30
python-dotenv==1.0.0
langgraph==0.0.19
pydantic==2.5.2
numpy==1.26.2
graphviz==0.20.1
